Corrected bool handling in SQLite and MySQL Grammars

+++ src/Illuminate/Database/Schema/Grammars/SQLiteGrammar.php
@@ -433,7 +433,7 @@ class SQLiteGrammar extends Grammar {
 	 */
 	protected function typeBoolean(Fluent $column)
 	{
-		return 'tinyint';
+		return 'integer';
 	}
 
 	/**
